Replace `tier` variable with `use_case` for Service Bus (#811)
Replace the `tier` variable with a new `use_case` variable for tiering
configuration. New value are:

- default

Update some variable validation method:

- `allowed_ips`: Now it can be defined only if `use_case` is `default`
- `subnet_pep_id`: Now it can be defined only when private_enpoint will
be created (in this case only when `use_case` is `default`)
- `private_dns_zone_resource_group_name`: As subnet pep id now it can be
defined only when `private_enpoint` will be created

Previously, validation for some of this variables was strictly tied to
specific `tier`.
This change decouples the validation from the `use_case` definitions,
making it easier to add new profiles in the future without having to
update multiple validation blocks.

Resolves: CES-1211